media-gfx/ufraw:        revbump to fix built
media-gfx/ufraw has been failing due to changes in exiv2 library. Based
on https://github.com/sergiomb2/ufraw/issues/12 I added a new patch.
Thanks to @Flexximilian for that!

NOTE: media-gfx/ufraw is not in official portage for 2+ years and it
shows a lot of warnings, so I cannot vouch for its stability/suitability;
I only occasionally use it.